Add files via upload
Code to perform a scan of the object present in front of the robot. A stepper motor mechanism was used to move the two IR sensors up and towards the side respectively. The motors stopped when the IR sensors were not able to detect anything in front of them.
This way the exact dimensions of the block were found out using IR sensors and stepper motors.